Add sideScreen component
